Separate source file and docs for AndroidHiltTransform
